Code refactoring to KarmaApp. Smooth camera movement
