+++ Karma/src/Karma.h
@@ -7,5 +7,14 @@
 #include "Karma/Log.h"
 #include "Karma/ImGui/ImGuiLayer.h"
 
+#include "Karma/Renderer/Shader.h"
+#include "Karma/Renderer/VertexArray.h"
+#include "Karma/Renderer/OrthographicCamera.h"
+#include "Karma/Renderer/Renderer.h"
+#include "Karma/Renderer/RenderCommand.h"
+#include "Karma/Renderer/Buffer.h"
+
+#include "Karma/Input.h"
+
 // Entry point from Karma engine
 #include "Karma/EntryPoint.h"
