+++ Karma/src/Karma/Application.cpp
@@ -9,7 +9,6 @@ namespace Karma
 	Application* Application::s_Instance = nullptr;
 
 	Application::Application()
-		: m_Camera(-1.6f, 1.6f, -0.9f, 0.9f)
 	{
 		KR_CORE_ASSERT(!s_Instance, "Application already exists!");
 		s_Instance = this;
@@ -18,121 +17,6 @@ namespace Karma
 
 		m_ImGuiLayer = new ImGuiLayer();
 		PushOverlay(m_ImGuiLayer);
-
-		m_VertexArray.reset(VertexArray::Create());
-		
-		float vertices[3 * 7] = {
-			-0.5f, -0.5f, 0.0f, 1.0f, 0.0f, 0.0f, 1.0f,
-			 0.5f, -0.5f, 0.0f, 0.0f, 1.0f, 0.0f, 1.0f,
-			 0.0f, 0.5f, 0.0f, 0.0f, 0.0f, 1.0f, 1.0f
-		};
-
-		std::shared_ptr<VertexBuffer> m_VertexBuffer;
-		m_VertexBuffer.reset(VertexBuffer::Create(vertices, sizeof(vertices)));
-		
-		{
-			BufferLayout layout = {
-				{ ShaderDataType::Float3, "a_Position" },
-				{ ShaderDataType::Float4, "a_Color"	}
-			};
-
-			m_VertexBuffer->SetLayout(layout);
-		}
-		
-		m_VertexArray->AddVertexBuffer(m_VertexBuffer);
-
-		uint32_t indices[3] = { 0, 1, 2 };
-
-		std::shared_ptr<IndexBuffer> m_IndexBuffer;
-		m_IndexBuffer.reset(IndexBuffer::Create(indices, sizeof(indices) / sizeof(uint32_t)));
-		m_VertexArray->SetIndexBuffer(m_IndexBuffer);
-
-		std::string vertexSrc = R"(
-			#version 330 core
-			
-			layout(location = 0) in vec3 a_Position;
-			layout(location = 1) in vec4 a_Color;
-
-			uniform mat4 u_ViewProjection;
-
-			out vec4 v_Color;
-			
-			void main()
-			{
-				gl_Position = u_ViewProjection * vec4(a_Position, 1.0f);
-				v_Color = a_Color;
-			}
-		)";
-
-		std::string fragmentSrc = R"(
-			#version 330 core
-			
-			layout(location = 0) out vec4 color;
-
-			in vec4 v_Color;
-
-			void main()
-			{
-				color = vec4(0.8, 0.2, 0.3, 1.0);
-				color = v_Color;
-			}
-		)";
-		m_Shader.reset(new Shader(vertexSrc, fragmentSrc));
-
-		// Drawing square
-		m_SquareVA.reset(VertexArray::Create());
-
-		float verticesBSQ[3 * 4] = {
-			-0.25f, -0.25f, 0.0f,
-			 0.25f, -0.25f, 0.0f,
-			 0.25f, 0.25f, 0.0f,
-			 -0.25f, 0.25f, 0.0f
-		};
-
-		std::shared_ptr<VertexBuffer> squareVB;
-		squareVB.reset(VertexBuffer::Create(verticesBSQ, sizeof(verticesBSQ)));
-
-		{
-			BufferLayout layout = {
-				{ ShaderDataType::Float3, "a_Position" },
-			};
-
-			squareVB->SetLayout(layout);
-		}
-
-		m_SquareVA->AddVertexBuffer(squareVB);
-
-		uint32_t indicesBSQ[6] = { 0, 1, 2, 2, 3, 0 };
-		std::shared_ptr<IndexBuffer> squareIB;
-		squareIB.reset(IndexBuffer::Create(indicesBSQ, sizeof(indicesBSQ) / sizeof(uint32_t)));
-		m_SquareVA->SetIndexBuffer(squareIB);
-
-
-		std::string vertexSrcBSQ = R"(
-			#version 330 core
-			
-			layout(location = 0) in vec3 a_Position;
-			
-			uniform mat4 u_ViewProjection;
-
-			void main()
-			{
-				gl_Position = u_ViewProjection * vec4(a_Position, 1.0f);
-			}
-		)";
-
-		std::string fragmentSrcBSQ = R"(
-			#version 330 core
-			
-			layout(location = 0) out vec4 color;
-
-			void main()
-			{
-				color = vec4(0.2, 0.3, 0.8, 1.0);
-			}
-		)";
-
-		m_BlueSQShader.reset(new Shader(vertexSrcBSQ, fragmentSrcBSQ));
 	}
 
 	Application::~Application()
@@ -145,18 +29,6 @@ namespace Karma
 	{
 		while (m_Running)
 		{
-			RenderCommand::SetClearColor({ 0.1f, 0.1f, 0.1f, 1 });
-			RenderCommand::Clear();
-			
-			m_Camera.SetPosition({camData.x_Pos, camData.y_Pos, 0.0f});
-			m_Camera.SetRotation(camData.angle);
-			Renderer::BeginScene(m_Camera);
-
-			Renderer::Submit(m_VertexArray, m_Shader);
-			Renderer::Submit(m_SquareVA, m_BlueSQShader);
-			
-			Renderer::EndScene();
-
 			// The range based for loop valid because we have implemented begin()
 			// and end() in LayerStack.h
 			for (auto layer : m_LayerStack)
@@ -205,33 +77,6 @@ namespace Karma
 			if (e.IsHandled())
 			{
 				break;
-			}
-		}
-
-		if (e.GetEventType() == EventType::KeyPressed)
-		{
-			KeyPressedEvent* kPressedEvent = static_cast<KeyPressedEvent*>(e.GetObjPointer());
-			int keycode = kPressedEvent->GetKeyCode();
-			switch (keycode)
-			{
-				case GLFW_KEY_A:
-					camData.x_Pos -= 0.1f;
-					break;
-				case GLFW_KEY_D: 
-					camData.x_Pos += 0.1f;
-					break;
-				case GLFW_KEY_W:
-					camData.y_Pos += 0.1f;
-					break;
-				case GLFW_KEY_S:
-					camData.y_Pos -= 0.1f;
-					break;
-				case GLFW_KEY_R:
-					camData.angle += 3.0f;
-					break;
-				case GLFW_KEY_T:
-					camData.angle -= 3.0f;
-					break;
 			}
 		}
 	}
