+++ Karma/src/Karma/Renderer/OrthographicCamera.cpp
@@ -3,6 +3,10 @@
 
 namespace Karma
 {
+	OrthographicCamera::OrthographicCamera()
+	{
+	}
+	
 	OrthographicCamera::OrthographicCamera(float left, float right, float bottom, float top)
 		: m_ProjectionMatrix(glm::ortho(left, right, bottom, top, -1.0f, 1.0f)), m_ViewMatrix(1.0f)
 	{
