+++ Karma/src/Karma/Renderer/OrthographicCamera.h
@@ -8,6 +8,7 @@ namespace Karma
 	class KARMA_API OrthographicCamera
 	{
 	public:
+		OrthographicCamera();
 		OrthographicCamera(float left, float right, float bottom, float top);
 
 		const glm::vec3& GetPosition() const { return m_Position; }
